Test listos.
Se realizaron los Test para usuarios.

+++ src/main/java/com/dangerbook/usuarios/dangerbook/usuarios/model/Usuario.java
@@ -14,7 +14,7 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario" )
-    private Long id_usuario;
+    private Integer id_usuario;
 
     @Column(name = "nombre", nullable = false, length = 100)
     private String nombre;
